+++ docs/source/changelog.rst
@@ -5,7 +5,11 @@ pycite's changelog
 **pycite 0.1.2**
 
 
-* Initial support for a script mode. 
+* 
+  For PubMed citations, an optional ``show_doi`` argument was added to control whether dois should exist in the citation.
+
+* 
+  Initial support for a script mode. 
 
 **pycite 0.1.1**